+++ src/main/java/org/truffleruby/interop/OutgoingForeignCallNode.java
@@ -76,22 +76,63 @@ public abstract class OutgoingForeignCallNode extends RubyBaseNode {
         return System.identityHashCode(receiver);
     }
 
-    @Specialization(guards = { "name == cachedName", "cachedName.equals(INDEX_READ)", "args.length == 1" }, limit = "1")
-    protected Object indexRead(
+    @Specialization(
+            guards = {
+                    "name == cachedName",
+                    "cachedName.equals(INDEX_READ)",
+                    "args.length == 1",
+                    "isBasicInteger(first(args))" },
+            limit = "1")
+    protected Object readArrayElement(
             Object receiver, String name, Object[] args,
             @Cached(value = "name", allowUncached = true) @Shared("name") String cachedName,
-            @Cached InteropNodes.ReadNode readNode) {
+            @Cached InteropNodes.ReadArrayElementNode readNode) {
         return readNode.execute(receiver, args[0]);
     }
 
     @Specialization(
-            guards = { "name == cachedName", "cachedName.equals(INDEX_WRITE)", "args.length == 2" },
+            guards = {
+                    "name == cachedName",
+                    "cachedName.equals(INDEX_READ)",
+                    "args.length == 1",
+                    "isRubySymbol(first(args)) || isRubyString(first(args))" },
             limit = "1")
-    protected Object indexWrite(
+    protected Object readMember(
             Object receiver, String name, Object[] args,
             @Cached(value = "name", allowUncached = true) @Shared("name") String cachedName,
-            @Cached InteropNodes.WriteNode writeNode) {
+            @Cached InteropNodes.ReadMemberNode readNode) {
+        return readNode.execute(receiver, args[0]);
+    }
+
+    @Specialization(
+            guards = {
+                    "name == cachedName",
+                    "cachedName.equals(INDEX_WRITE)",
+                    "args.length == 2",
+                    "isBasicInteger(first(args))" },
+            limit = "1")
+    protected Object writeArrayElement(Object receiver, String name, Object[] args,
+            @Cached(value = "name", allowUncached = true) @Shared("name") String cachedName,
+            @Cached InteropNodes.WriteArrayElementNode writeNode) {
+
         return writeNode.execute(receiver, args[0], args[1]);
+    }
+
+    @Specialization(
+            guards = {
+                    "name == cachedName",
+                    "cachedName.equals(INDEX_WRITE)",
+                    "args.length == 2",
+                    "isRubySymbol(first(args)) || isRubyString(first(args))" },
+            limit = "1")
+    protected Object writeMember(Object receiver, String name, Object[] args,
+            @Cached(value = "name", allowUncached = true) @Shared("name") String cachedName,
+            @Cached InteropNodes.WriteMemberNode writeNode) {
+        return writeNode.execute(receiver, args[0], args[1]);
+    }
+
+    protected static Object first(Object[] args) {
+        return args[0];
     }
 
     @Specialization(guards = { "name == cachedName", "cachedName.equals(CALL)" }, limit = "1")
@@ -213,7 +254,7 @@ public abstract class OutgoingForeignCallNode extends RubyBaseNode {
             case TO_ARY:
                 return "to_array";
             case SIZE:
-                return "size";
+                return "array_size";
             case KEYS:
                 return "keys";
             case RESPOND_TO: