+++ test/truffle/compiler/pe/interop/interop_pe.rb
@@ -17,5 +17,5 @@ end
 
 
 # The Truffle API's these use are different from GraalVM 0.12 and Truffle 0.13
-tagged example "Truffle::Interop.read(InteropPE::A.new(42), :@foo)", 42
-tagged example "Truffle::Interop.read(InteropPE::A.new(42), '@foo')", 42
+tagged example "Truffle::Interop.read_member(InteropPE::A.new(42), :@foo)", 42
+tagged example "Truffle::Interop.read_member(InteropPE::A.new(42), '@foo')", 42